bin: Migrate swh-hashtree to swh-model
This is a temporary move.  This will eventually converge with
swh.model.cli as the directory hashing ability overlaps.  What's
missing in the cli for now is the ability to filter patterns.

This is mostly used to compute hashes on tree coming from multiple
dvcs. This imposes the need to be able to filter pattern like .svn,
.hg, etc...