Add to_model() method to from_disk.{Content,Directory}, to convert to canonical model objects.
They will be used by loaders, so they can deal only with
model objects, instead of having to do the same conversion themselves.

This removes the `data` and `save_path` arguments of `from_file` and
`from_disk`, as data loading is always deferred from now on.
To access it, users are now expected to either open the data files
themselves, or us `.to_model().with_data()`.