Fix tests of special devices.
Regular files were created, as the 'mode' argument of os.mknod was missing.
However, creating devices requires root; so we can't reasonably do that in
tests.

Instead, we're using /dev/null instead of creating one.
And while we're at it, let's also use /dev/zero (which, if not handled
properly, will result in an infinite read).