tests: Clean hashutil._blake2_hash_cache after mocking blake2 functions.
Depending on the order in which tests are run, these tests may insert
lambdas with mocked blake2 functions in their closure to be inserted in
hashutil._blake2_hash_cache; causing all future tests to fail.

While this does not happen with the default order of tests, it does when
using pytest-xdist.