identifiers: Expose manifest computation
Before this commit, manifests were only computed internally before
hashing, so they were not available to outside modules.

This makes testing the module very painful, because identifier
functions can only be tested by checking the hash; so test failures
did not show mismatches between the computed manifest and
the expected one.

Additionally, the 'git bare cooker' of the vault is likely to use
these as well, as it needs to format git objects in the same format.