Move manifest computation functions from identifiers.py to git_objects.py
Since they are used by the vault for non-identifier-related stuff,
I think it makes sense to move them to a new module.

identifiers.py is now an empty shell, as all its features were moved to
other modules and it only contains reexports and backward-compat
functions. Therefore, it should be considered deprecated from now on.