TimestampWithTimezone: Make 'offset_bytes' required and remove 'negative_utc'.
'offset' becomes a property instead of an attribute and constructor argument.

This also removes both from the output of `.to_dict()`.

This is step 6 of https://forge.softwareheritage.org/T3752

This will break packages that still use the constructor directly,
ie. swh-storage and swh-loader-git (and tests of swh-loader-core
and swh-loader-svn)