Revert "Restore 'offset' and 'negative_utc' arguments and make them optional"
This reverts commit a0f5436273ef6f5b62a388ae131ed5afa7287d00.

This means this commit removes the 'offset' and 'negative_utc'
arguments.

It also removes the 'negative_utc' attribute (which is not used
anymore), but keeps an 'offset' property, which is an alias to
'offset_minutes()'. This is only to keep this commit readable;
the next commit will remove this alias.