Add parameter 'ignore_unresolved' to snapshot_git_object
It allows building a snapshot manifest despite some branches being
unresolved (instead of raising an error).

This feature was removed in 57ae405d312879bec19107d29a20c2c290d7861d
but in the end, Snapshot.compute_hash() will need to use it so that
swh-scrubber does not crash when checksum such snapshots.