model: directly implement `to_dict` method for model class
The current implementation is very generic and creates many intermediate
dictionary and looping. This is called a lot by the RPC so having
something fast would be useful. In that spirit we move to a much more
verbose but direct approach to defining these.

This greatly improve the performance of the loader with and without
pypy

For example, on a run of the git loader on the "git" loader:

- cpython with a small buffer: -13% user time (1640s → 1425s)
- cpython with a large buffer: -25% user time (940s → 670s)
- pypy with a large buffer: -15% user time (530s → 450s)
- pypy with a small buffer: -25% user time (1100s → 800s)
                            (about ½ of the rpc overhead)