Make Content.unique_key a dict of hashes
This will allow properly storing contents with SHA1 hash collisions in
swh.journal.

This is a breaking change, but the only concrete effect is that it will
prevent kafka from deduplicating the same content if it is written
before and after the change. We expect this to have very low impact, as
in practice all objects are checked for existence in storage before
being sent on to the journal, so duplicates are expected to be very
close to one another in time.